Refactored main.py and added csv to excel converter (works)
Most of main.py functions and code have been moved to a new file called handler.py under the class Analysis. An excel converter was created to transform the csv file to an Excel file format. It is in a new class called simply Excel. Thus, enabling formulas and the like with the file.

- The output excel file still needs to have its path updated. It saves to the directory main.py is contained in.